Add editor for gameobject selection
Also moved some classes to other files.
Next up is seeing if we can componentize these two parts of the editor -- blendshapes and gameobjects. Since there will be multiple animationsets, including the default, we can't just hardcode stuff!
Then we can refine the editor stuff for these to be able to set the blendshape value and whether or not stuff is active.