Was attempting the path of abstraction instead of UTF
From a ports-and-adapters-perspective, it was going fine for driven ports, since we can get useful testing out of abstracts when outputting generation, so to speak.

But for driver ports, the useful stuff comes with interacting with gameobjects and components (aka shapekeys of a skinnedmeshrenderer), and abstracting all of that out is both annoying, leaves a lot of adapter implementation untested, and covering that would require a lot of integration testing that would have to be done in UTF anyway.

So will prob pivot and go for UTF!