Started working on the next test, and finally remembered where i left off
Was about to start having the generator mess with assets, but finally remembered the strat to not do that, which is kinda nice. We'll return a list of animations from the generation so that the driver controller can turn them into assets. We'll also edit the reference artifacts in place and rely on the driver controller to generate asset copies before passing them in.