Got working unit tests that can just about start testing generation
Aka mainly have VrcAssets keeping track of old and new things by being able to clone the current and record it as original, which can then be compared.

There was some friction that was worked out around messing with components and objects. It's worth reconsidering having the generator touch components, though having components touch generator, for the sake of having a generate button and whatnot, makes sense in that case. Is obnoxious to have two sets of the same config, tho. However, this may not be necessary with right editor code and simply serializing the editor-based config into the component.