Add test and fix bug around bool condition thresholds; refactor a bit
Bug was found while reading

Refactor is simply around referring to processed states' indexes, rather than corresponding parameter value. It more accurately conveys the intent to base things off the ordering of parameters. While not as plainly visible in code, the tests convey that the reference animation can be considered index0 among the set of animations, even though not in that list.