Refactor animation/clip management into the newer code from last time
Also added a test that ensures the generation result only returns the clips that are used, which comes into play if state management is off and not all states are present. Previously, we'd return unused clips, as well. Arguably it could only be a good thing to save extra clips, though.