Partway thru refactoring, but will need to experiment
Basically ProcessedAnimation is looking so close to AnimationConfiguration that it might be time to merge them. AnimationConfiguration isn't a component. It is, however, a serializable class with lack of encapsulation. But it's not like we can't do that with unity serialization.

So let's see if we can't unify the two and let them be shared. The only risk is that they kinda serve two different purposes, but it doesn't have to be that way. The purpose of the "component-side" class is to be converted to the processed class, after all.