Rearranged files and moved AnimationConfiguration back to components
components needs to reference this ofc. but, as it turns out, editor needs to ofc reference components for a few reasons, including that the inspector guis will go in editor. so we got a circular reference we gotta fix, hence moving it back. the methods had to come out because they reference editor code, and they now sit in an AnimationConfigurationHandler class.

Also took another quick look at adding custom components to avatars. We can strip them at avatar build time, but the sdk still errors out no matter what, even with EditorOnly tag. Whenever we get around to doing per-avatar configs, we'll need a bit of indirection to get around the limitation. Or we can do some hacky stuff to make our components allowed, but let's not do that.

A really funny hack would be to store our data in meshes containing materials with particular settings -- that can be stripped out later.

Anyway, the workaround is almost certainly to have a component to make configs specific to an avatar, and all siblings and children fall under that scope.