Temporarily have a version that cleanly adds and remove items
But we'll revert it right after!
The thing is, later on, we'll be *inserting* animatables into various places. Mainly, this will be updating other animations when the reference animation is modified. This could also be for reordering the reference animation, though we won't support it anytime soon. In such a scenario, the whole reconfigure/rebind workflow works a lot better.

On that note, some initial thoughts on how to implement that. Let's try events -- the uielements kind!
If an animation configuration gets a delete event for a blend shape or a gameobject, it triggers a delete that sends no events. Add will work much the same.

And for the refresh, there's more work to do. We need to deserialize the smrs, group them, then order the shape keys by the indexes they have in the smr.

Oh and I just remembered an idea to try that I forgot. SerializedObject.Update! Could experiment with manipulating the objects directly. (also note that this is theoretically called automatically but who knows)
Though the really iffy thing is undos, which would need to be done manually. The really iffy part is if an undo will update the ui or not