Can now add animations! And now support undos of array changes!
The array change undos thing was particularly tricky. The only hypothetical way I could find was to hook into the undoredo callback and just refresh. This created a new problem where we might try to refresh something that was removed. Catching the resulting exception worked, but it littered the logs. Found a particularly hacky way of finding out if the array was deleted or not.

Also removed the canEdit/canRemove parameters to simplify the design. Instead, we read the now-renamed isReferenceAnimation to decide to hide or unhide certain elements from the UI.