Have basically added blend shape grouping!
Still lots of cleanup to do

And there's a bug! We're using the passed-in index to decide whether to add or reuse elements. We can't use this overall index because we do this whole grouping thing. We could track indexes externally, but it'll be roughly the same amount of work to pick a better implementation, so we'll do that.