After great pain and suffering, got reference match checking working
Will need lots of redesign, though, since things are getting complicated. While we still need a picker to finish up the blendshape support, this is the main hard piece that was left out.

Also did a refactor of serializedpropertycontainer to not be generic, since it didn't need to be anymore.

Next up will be some redesigning! No thoughts yet, though. The main pain points are due to the excessive passing of data round and round.