A bit of refactoring
First, AnimatableBlendShapeField doesn't need a rebind because it's just bound once. Instead, do it in the ctor.

Simplified the reference matching stuff a bit, though still not simple enough.

Appear to have fixed a bug, since we dont get NREs anymore, but adding and removing animations still doesnt quite work. The UI goes weird, but reloading it looks fine, so probably just a ui problem and not a serialization problem.