Fixed one of what is actually two bugs in adding animations
The blend shape container renderer doesn't actually support reuse of the element, since we do caching inside. Therefore, we now clear it on rebind! We also check to see if it's new so we don't go thru all the rebinding steps if we don't need to.

Still, new animations ui still looks weird, so not sure about that yet.