Fixed a bug where we couldn't remove last blend shape from ui
This was introduced when having the SerializedPropertyContainer not refresh if there were no elements, which was definitely wrong in hindsight. This matters for all renderers, as well.

Though it certainly makes sense to not finalize if initialization isn't run, the real design issue is that we handled initialization by seeing if we were processing the first element. The fix, in this case, is to have an initialization method in the interface if we have a finalize one.