Got a working technique to maintain asset ids
Also ends up being pretty performant, quicker than prior attempts, meaning it's probably that we were able to take out the asset copy that helps.

The technique is to copy in-memory via EditorUtility.CopySerialized for generation. Then, when ready to save, we load the previously generated one and copy the working object over to the previously generated one.