Fixed a bug where saving to old generated assets didnt work for menus
It was recognized the code shouldn't have worked before, and it didn't! Didn't test subsequent generations.

The general idea is that we have all the data needed to map working menu to generated menu, so we go through each menu and swap in the working menu to the generated menu.

The current implementation is also a bit more complicated than it needs to be because a previous version was a bit more complicated than it needed to be. We basically generated a list from leaves to trunk and fix in that order. This was probably actually the wrong order, and the order no longer matters anyway. Still gotta clean it up tho, and wanna do it in a separate commit.

Also lumped in some quick uxml filex, as well, though they don't really seem to impact behavior.