Did a redesign to get rid of ConfigOperations
We now have a slimmer AnimatorLayerConfigurator, and much of the ConfigOperations stuff now exists in fields themselves, which are exposed by this configurator class. The original class was getting pretty dense, and the design smelled because it became the place to put implementation not meant to be used internally, in addition to its stated purpose of being the port exposed for unit testing.

As a side-note, we're doing away with the `hide on reference match` feature. While definitely useful, I think it'll become confusing as our config gets more complicated. We can always add it back later!

Not 100% sure but also probably fixed a long-standing bug where, if the default animation is changed, followed by a name change, the name change wouldn't reflect until inspector reload. It didn't look like we were propagating the information through. It was fixed by adding a hidden field that ends up taking care of all the plumbing.