Fix undo/redo for default and toggle off animation popups.
It probably never worked quite right. The underlying animation list would never get the right set on undo/redo. Now, instead of adding and removing to/from this list, we pull the list of animations and apply it to the popup list. This refresh happens on undo/redo, as well as when animations are added or removed.

Also, while investigating, change the implementation of SerializedProperty removal. If it ends up not working in all cases, it could be changed back. The current implementation doesn't require reversing the set; while not a big deal, it's technically better this way so keeping.