Do a pass on all existing unit tests
They all work, and they all make sense given the new design, including test names. This is also to say we technically fixed bugs!

Some small design changes were done/clarified. In particular, we no longer add animator controller layers if we're not managing states, where we used to before. We also log warnings if we're unable to find the states to which we want to add motions to.